♻️ Introduce loaders.ts: split request facing reads from the DAL
Session scoped reads (getCurrentUser, getActiveSpace, getSeatUsage) move
to features/<domain>/loaders.ts; data.ts keeps only parameterized reads
that trust their input. Loaders never import @rallly/database and
delegate to data.ts, so query logic cannot fork between the session path
and the API path.

The api-keys page state reads dissolve into their single consumer route
per the route-by-default placement rule; the gate for the api-key
actions is inlined into requireApiAccess.

Lint: the enumerated session scoped module ban in API routes becomes the
glob @/features/**/loaders; data.ts and mutations.ts are banned from
next/headers and next/navigation.
