+++ CLAUDE.md
@@ -179,7 +179,8 @@ Always use gitmoji prefixes in commit messages. Follow the gitmoji convention (h
 
 **Feature file vocabulary** — closed set; new file names require a team decision. Applies recursively in sub-concern directories:
 
-- `data.ts` — reads (Prisma queries), must start with `import "server-only"`
+- `data.ts` — parameterized reads (Prisma queries), must start with `import "server-only"`. Trusts its input: every query takes explicit arguments and carries its tenant scope in the where clause. Never reads the request — no `next/headers`, `next/navigation`, or session state (lint enforced)
+- `loaders.ts` — request facing reads: resolve the actor from the session, apply page semantics (redirects to /login and /setup, `InvalidSessionError` on bans), and delegate to `data.ts` with proven scope. Loaders always consume `data.ts`, never the reverse, and never import `@rallly/database` — so query logic cannot fork between the session path and the API path. Server components call loaders; API routes, webhooks and cron call `data.ts` with proven scope from their own gate (API routes are lint banned from `@/features/**/loaders`). Naming note: `session-data.ts` and `queries.ts` were considered and rejected
 - `mutations.ts` — writes + cache invalidation, must start with `import "server-only"`. Trusts its input: takes explicit parameters (`userId`, not headers), no session reads, no `headers()`, callable from system contexts (webhooks, cron, moderation). Authorization happens in `actions.ts` (safe-action middleware + CASL against the database). Never call Better-Auth endpoints from a mutation — they resolve the target user from request headers and authorize against the session snapshot; use adapter-level APIs (`authLib.$context` → `internalAdapter`) instead
 - `actions.ts` — `"use server"` actions, thin wrappers over mutations, validated via safe-action + `schema.ts`. Owns authentication and authorization (safe-action clients + CASL checks against database state, never against the session snapshot). Exception to the thin-wrapper rule: writes whose target user is defined by the session (self-profile updates) call the Better-Auth endpoint directly in the action — the endpoint refreshes the session snapshot and cookie cache in one step. Rule of thumb: target user from a parameter → mutation via `internalAdapter`; target user from the session → Better-Auth endpoint in the action
 - `schema.ts` — Zod schemas (isomorphic, no Prisma imports)
@@ -193,7 +194,16 @@ Always use gitmoji prefixes in commit messages. Follow the gitmoji convention (h
 - `assets/` — static files used by the feature
 - NOT allowed: `index.ts` barrels, `helpers.ts`, `queries.ts`, `hooks.ts`, `lib/`, `libs/`
 
-**DAL enforcement** — `@rallly/database` may only be imported from `features/**/data.ts` and `features/**/mutations.ts` (lint-enforced via `noRestrictedImports`; existing violations sit on a shrinking migration allowlist in `apps/web/biome.json`). Parameterized reads take their tenant scope as `spaceId: AuthorizedSpaceId` (from `@/features/space/types`); only the session gate (`createSpaceDTO`) and the API key middleware may cast to it. API routes under `app/api/**` (except tRPC and better-auth) must not import session scoped DAL modules — they authenticate their own way and pass proven scope to parameterized reads. `api/private` handlers own serialization and must parse response bodies through their zod schemas.
+**Read/write symmetry** — the request facing / core split is the same on both sides:
+
+|  | Request facing (session, redirects, auth) | Core (parameterized, trusts input) |
+| -- | -- | -- |
+| Reads | `loaders.ts` | `data.ts` |
+| Writes | `actions.ts` | `mutations.ts` |
+
+**Loader placement** — route by default, promote on the second consumer. Page data needs are view shaped, so the server component composes in the route: it calls the session gate loaders (`getActiveSpace()`, `getCurrentUser()`) and passes proven scope to `data.ts` reads; extract to a route private file only when the page gets long. Never import another route's loader. Promote to `features/<domain>/loaders.ts` only when a second route segment needs the same read — that is evidence of a domain concept, not a view shape. The test: a loader belongs in the feature when its contract is stated in domain terms and every consumer wants exactly that contract; it stays in the route when its shape is "what this page happens to display." Most features never have a `loaders.ts` (today: `space`, `user`).
+
+**DAL enforcement** — `@rallly/database` may only be imported from `features/**/data.ts` and `features/**/mutations.ts` (lint-enforced via `noRestrictedImports`; existing violations sit on a shrinking migration allowlist in `apps/web/biome.json`; `loaders.ts` is absent from the allowlist, so database imports are banned there by default). Parameterized reads take their tenant scope as `spaceId: AuthorizedSpaceId` (from `@/features/space/types`); only the session gate (`createSpaceDTO`) and the API key middleware may cast to it. API routes under `app/api/**` (except tRPC and better-auth) must not import `@/features/**/loaders` — they authenticate their own way and pass proven scope to parameterized reads. `api/private` handlers own serialization and must parse response bodies through their zod schemas.
 
 **Cross-feature imports** — allowed, public surface only (the vocabulary files above); never reach into another feature's internals. No cycles between features (CI-enforced). For UI, prefer composing multiple features at the page level in `app/` over feature-to-feature component imports.
