+++ apps/web/src/app/[locale]/(space)/layout.tsx
@@ -1,7 +1,7 @@
 import { SessionRefresher } from "@/components/session-refresher";
 import { PayWall } from "@/features/billing/components/pay-wall";
 import { SpaceProvider } from "@/features/space/client";
-import { getActiveSpace } from "@/features/space/data";
+import { getActiveSpace } from "@/features/space/loaders";
 import { UserProvider } from "@/features/user/client";
 import { TimeZoneMismatchDialog } from "@/features/user/components/timezone-mismatch-dialog";
 import { getLocale } from "@/i18n/server/get-locale";
