+++ apps/web/src/app/[locale]/(space)/settings/api-keys/components/api-keys-list.tsx
@@ -15,10 +15,7 @@ import { Trans } from "@/i18n/client";
 import { useDateTime } from "@/lib/datetime/client";
 import { RevokeApiKeyButton } from "./revoke-api-key-button";
 
-type ApiKey = Extract<
-  Awaited<ReturnType<typeof getSpaceApiKeys>>,
-  { ok: true }
->["apiKeys"][number];
+type ApiKey = Awaited<ReturnType<typeof getSpaceApiKeys>>[number];
 
 export function ApiKeysList({ apiKeys }: { apiKeys: ApiKey[] }) {
   const { toRelativeTime } = useDateTime();
