+++ apps/web/src/app/[locale]/admin-setup/page.tsx
@@ -11,7 +11,7 @@ import {
   EmptyStateTitle,
 } from "@/components/empty-state";
 import { isInitialAdmin } from "@/features/instance-settings/utils";
-import { getCurrentUser } from "@/features/user/data";
+import { getCurrentUser } from "@/features/user/loaders";
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
 import { buildSafeRedirectUrl } from "@/lib/utils/redirect";
