+++ apps/web/src/features/user/data.ts
@@ -2,10 +2,7 @@ import "server-only";
 
 import type { User } from "@rallly/database";
 import { prisma } from "@rallly/database";
-import { cache } from "react";
 import type { UserDTO } from "@/features/user/schema";
-import { getSession } from "@/lib/auth";
-import { InvalidSessionError } from "@/lib/errors/invalid-session-error";
 
 export const createUserDTO = (user: User): UserDTO => ({
   id: user.id,
@@ -23,29 +20,17 @@ export const createUserDTO = (user: User): UserDTO => ({
   deletedAt: user.deletedAt ?? undefined,
 });
 
-/**
- * The current signed-in user, fetched from the database. Returns null
- * when there is no session or the user is a guest — the caller decides
- * how to respond (redirect, 401, etc.). Throws InvalidSessionError when
- * the session references a user that no longer exists or is banned.
- */
-export const getCurrentUser = cache(async () => {
-  const session = await getSession();
+export const getUser = async (userId: string) => {
+  const user = await prisma.user.findUnique({
+    where: { id: userId },
+  });
 
-  if (!session?.user || session.user.isGuest) {
+  if (!user) {
     return null;
   }
 
-  const user = await prisma.user.findUnique({
-    where: { id: session.user.id },
-  });
-
-  if (!user || user.banned) {
-    throw new InvalidSessionError();
-  }
-
   return createUserDTO(user);
-});
+};
 
 export function getUserProfile(userId: string) {
   return prisma.user.findUnique({
